NP-ikke-logg-poster-med-cristinId, bugg fixed
