+++ src/main/java/no/sikt/nva/scrapers/DublinCoreScraper.java
@@ -13,6 +13,7 @@ import no.sikt.nva.channelregister.ChannelRegister;
 import no.sikt.nva.exceptions.DublinCoreException;
 import no.sikt.nva.model.BrageLocation;
 import no.sikt.nva.model.ErrorDetails;
+import no.sikt.nva.model.NvaType;
 import no.sikt.nva.model.WarningDetails;
 import no.sikt.nva.model.WarningDetails.Warning;
 import no.sikt.nva.model.dublincore.DcValue;
@@ -23,7 +24,6 @@ import no.sikt.nva.model.record.Publication;
 import no.sikt.nva.model.record.PublishedDate;
 import no.sikt.nva.model.record.Record;
 import no.sikt.nva.model.record.Type;
-import no.sikt.nva.scrapers.TypeMapper.NvaType;
 import no.sikt.nva.validators.DoiValidator;
 import no.sikt.nva.validators.DublinCoreValidator;
 import nva.commons.core.JacocoGenerated;
