+++ src/test/java/no/sikt/nva/scrapers/TypeMapperTest.java
@@ -4,7 +4,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import java.util.List;
-import no.sikt.nva.scrapers.TypeMapper.NvaType;
+import no.sikt.nva.model.NvaType;
 import org.junit.jupiter.api.Test;
 
 public class TypeMapperTest {
